fog/requirements.txt
aiomqtt==2.3.0
asyncpg==0.29.0
flask==3.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
